+++ lib/display/select.d.ts
@@ -1,5 +1,7 @@
 import m, { ClassComponent, CVnode } from "mithril";
-import { IPropWidget } from "../interface/widget";
-export declare class SelectText implements ClassComponent<IPropWidget> {
-    view({ attrs: { field, value } }: CVnode<IPropWidget>): m.Vnode<any, any>;
+import { IOptionField, IPropWidget } from "../interface/widget";
+declare type TSelectWidget = IPropWidget<IOptionField>;
+export declare class SelectText implements ClassComponent<TSelectWidget> {
+    view({ attrs: { field, value } }: CVnode<TSelectWidget>): m.Vnode<any, any>;
 }
+export {};