+++ lib/input/currencyInput.d.ts
@@ -2,7 +2,7 @@ import m, { ClassComponent, CVnode } from "mithril";
 import { IConfig } from "../interface/config";
 import { IPropWidget, TProp, TPropStream } from "../interface/widget";
 export declare class CurrencyInput implements ClassComponent<IPropWidget> {
-    view({ attrs }: CVnode<IPropWidget>): m.Vnode<import("../interface/widget").IPropLayoutWidget, unknown>;
+    view({ attrs }: CVnode<IPropWidget>): m.Vnode<import("../interface/widget").IPropLayoutWidget<import("../interface/widget").IField>, unknown>;
 }
 export declare function formatCurrency(unitTotal: number, negativeStyle: IConfig["negativeStyle"]): string | undefined;
 export declare function propToNumber(value: TProp): number;