+++ lib/input/textarea.d.ts
@@ -3,5 +3,5 @@ import { IPropWidget } from "../interface/widget";
 import { ValidationBase } from "../validationBase";
 export declare class TextareaInput extends ValidationBase {
     protected readonly selector = "textarea";
-    view({ attrs }: CVnode<IPropWidget>): m.Vnode<import("../interface/widget").IPropLayoutWidget, unknown>;
+    view({ attrs }: CVnode<IPropWidget>): m.Vnode<import("../interface/widget").IPropLayoutWidget<import("../interface/widget").IField>, unknown>;
 }